homework added

+++ week1/3_input.py
@@ -10,9 +10,10 @@ print "Hi", name
 #input integers
 #ask for cost price. sell price is 1 more than cost price
 cost_input = raw_input("what is the cost price? ")
-cost_price = float(cost_input)
+cost_price = int(cost_input)
 
 sell_price = cost_price + 1
+#now change it so the sell price is 110% of cost price and is in decimal
 
 print "Cost Price", cost_price
 print "Sell Price", sell_price
